MC-1: review

+++ client/engine/renderable/tsx/_genetic/abstractTsxDOMRenderer.ts
@@ -3,7 +3,7 @@ import {VirtualCommentNode, VirtualNode} from "@engine/renderable/tsx/_genetic/v
 import {Optional} from "@engine/core/declarations";
 import {IRealNode} from "@engine/renderable/tsx/_genetic/realNode";
 import {AbstractElementCreator} from "@engine/renderable/tsx/_genetic/abstractElementCreator";
-import {VEngineTsxFactory} from "@engine/renderable/tsx/_genetic/vEngineTsxFactory.h";
+import {getComponentUuid, VEngineTsxFactory} from "@engine/renderable/tsx/_genetic/vEngineTsxFactory.h";
 import {BaseTsxComponent} from "@engine/renderable/tsx/base/baseTsxComponent";
 import {VirtualFragment} from "@engine/renderable/tsx/_genetic/virtualFragment";
 
@@ -36,6 +36,15 @@ export abstract class AbstractTsxDOMRenderer<T extends IRealNode> {
         return newVirtualDom;
     }
 
+    private needToBeDestroyed(oldNode:VirtualNode,newChildren:VirtualNode[]) {
+        if (!newChildren) return true;
+        const uuid = getComponentUuid(oldNode.props);
+        for (const ch of newChildren) {
+            if (uuid===getComponentUuid(ch.props)) return true;
+        }
+        return false;
+    }
+
     private removeNode(node:T,vNode:VirtualNode):void {
         node.removeSelf();
         VEngineTsxFactory.destroyElement(vNode);
@@ -45,7 +54,9 @@ export abstract class AbstractTsxDOMRenderer<T extends IRealNode> {
         const newNode = this.elementCreator.createElementByTagName(newVirtualNode);
         this.setGenericProps(newNode,newVirtualNode,parent);
         parent.replaceChild(node,newNode);
-        VEngineTsxFactory.destroyElement(oldVirtualNode);
+        if (this.needToBeDestroyed(oldVirtualNode,newVirtualNode.children)) {
+            VEngineTsxFactory.destroyElement(oldVirtualNode);
+        }
         return newNode;
     }
